chore: rename api_discovery > apidiscovery
