+++ internal/agent/apidiscovery/get_api_info.go
@@ -1,4 +1,4 @@
-package api_discovery
+package apidiscovery
 
 import (
 	"reflect"
