+++ internal/agent/apidiscovery/get_api_info_test.go
@@ -1,4 +1,4 @@
-package api_discovery
+package apidiscovery
 
 import (
 	"testing"
