+++ internal/agent/apidiscovery/get_body_data_type_test.go
@@ -1,4 +1,4 @@
-package api_discovery
+package apidiscovery
 
 import (
 	"testing"
