+++ internal/agent/apidiscovery/get_data_schema_test.go
@@ -1,4 +1,4 @@
-package api_discovery
+package apidiscovery
 
 import (
 	"encoding/json"
