+++ internal/agent/apidiscovery/is_json_content_type.go
@@ -1,4 +1,4 @@
-package api_discovery
+package apidiscovery
 
 import (
 	"strings"
