+++ internal/agent/apidiscovery/merge_api_auth_types_test.go
@@ -1,4 +1,4 @@
-package api_discovery
+package apidiscovery
 
 import (
 	"reflect"
