+++ internal/agent/apidiscovery/merge_data_schemas_test.go
@@ -1,4 +1,4 @@
-package api_discovery
+package apidiscovery
 
 import (
 	"testing"
